// src/app/modules/hcm/shared/models/laborRelationship/salariesforpayrolldata.ts
export class SalariesForPayrollData {

    idLaborRelationship: number = -1;
    salaryForLaborRelationShipId: number = -1;
    typeSalaryId: number = 1;
    typeSalary: string = "";   
    amount: number = 0;
    motiveId: number = 4;
	motive: string = "";
    validityDate: string = "1900-01-01";
    createdByUser: string = "";
    modifiedByUser: string = "";
    idCurrency: number = -1;                                                           
    currencySymbol: string = "";  
    abbreviation: string = "";  
    currency: string = "";    
    // createdById: number = -1;
    // modifiedById: number = -1;
    
}